host/requirements.txt
pyserial==3.5
python-dotenv==0.21.0
